refactor: extract parseSharedBeys, getComboStats, getComboName, STAT_DEFS to comboUtils
